# LibGfx/ICC+LibPDF: Speedups, round 2