Adds a dedicated matrix->matrix path. RGB profiles are usually matrix profiles, so this speeds up converting from most RGB colorspaces to sRGB (which is currently what LibPDF always converts to, but even if we make it support showing wide gamut colors on a wide gamut monitor, it'll still use a different RGB output profile) – that is, this is the common case.

Here are a few numbers from my local measurements. I timed four different workloads, at 5 different commits.

Workloads:

1. `time Build/lagom/bin/pdf --page 3 --render out.png ~/Downloads/0000/0000849.pdf`, ran a few times, eyeballed the min
2. `time Build/lagom/bin/pdf --debugging-stats ~/Downloads/0000/0000849.pdf`
3. `time Meta/test_pdf.py ~/Downloads/0000`
4. `time Meta/test_pdf.py ~/Downloads/0000` but with 0000849.pdf filtered out by adding `files = [f for f in files if '0000849' not in f]` to test_pdf.py locally

3 and 4 time rendering all pages in my 1000-file 0000.zip PDF/A test set (https://pdfa.org/new-large-scale-pdf-corpus-now-publicly-available/) on all cores in parallel. This time is dominated by rendering `0000849.pdf` (a file with many high-res JPEGs on 133 pages), and since the parallelism is limited to the pdf level, that file ends up serializing `test_pdf.py` – hence the run with that excluded, and the measurements on just that file.

The commits I measured at:
* "before icc": `64ec38e7fe1cb3a27ab35775459ea882e9548b4d^` (before the final commit in #22650)
* "before opts": `64ec38e7fe1cb3a27ab35775459ea882e9548b4d` (after the final commit in #22650)
* "master": 340936ad09b37d11583c92aa895ee22771208791 (what I currently have checked out; includes #22683)
* "icc-fasterer": my local branch for this PR
* "icc-fastererer": another local branch that has some more micro-optimizations

Data:

```
before icc    : 0000849.pdf p3: 0.65s, 0000849.pdf all: 10.9s, test_pdf.py all 43.4s, test_pdf.py all without 0000849.pdf: 42.9s
before opts   : 0000849.pdf p3: 1.84s, 0000849.pdf all: 1m17s, test_pdf.py all 1m57s, test_pdf.py all without 0000849.pdf: 1m07s
master        : 0000849.pdf p3: 1.33s, 0000849.pdf all: 51.4s, test_pdf.py all 1m25s, test_pdf.py all without 0000849.pdf: 56.9s
icc-fasterer  : 0000849.pdf p3: 1.13s, 0000849.pdf all: 41.2s, test_pdf.py all 1m13s, test_pdf.py all without 0000849.pdf: 53.9s
icc-fastererer: 0000849.pdf p3: 0.99s, 0000849.pdf all: 29.1s, test_pdf.py all   54s, test_pdf.py all without 0000849.pdf: 43.9s
```

This PR reduces time for p3 of 0000849.pdf from ~1.3s to ~1.1s (~15% speedup), is a ~19% speedup on 0000849.pdf for all pages, etc.